Added CordovaLocalWebServerStartOnSimulator setting, which defaults to true.
If set to false, when the user is running on the Simulator, we don't start the local web server. This is primarily to facilitate testing the wkwebview-engine plugin, which has a bug where file urls don't load on device (which is by design, there is a new loadFileURL API function, unreleased at this time) --  but file urls do load on Simulator.